update unitprice for warehouse

+++ ELIXIRETD.DATA/DATA ACCESS LAYER/DTOs/INVENTORY_DTO/MRP/DtoMRP.cs
@@ -30,7 +30,7 @@ namespace ELIXIRETD.DATA.DATA_ACCESS_LAYER.DTOs.INVENTORY_DTO.MRP
 
         public decimal SuggestedPo { get; set; }
         public decimal AverageIssuance { get; set; }
-        public decimal DaysLevel { get; set; }
+        public int DaysLevel { get; set; }
 
 
         public decimal ReserveUsage { get; set; }
